DarkJesus Collab to Add Custom Mobs to Bet
In this one, we added a custom iron golem that is super hard to beat and only drops an enchanted wood sword rarely. 

In addition, we propose giving the dead and ruined town of NSR a bunch of ghost witches that are invisible, slow, and spooky. They drop 10 netherwart upon dying that have the lore newfriend souls. In addition, there can only be one witch spawned per 2 minutes, so you'd have to stick around that town for a while to see it pop up (40% chance of spawning per 2 minutes).

This is partially a test of if the spawner option works. It is setting a hard location to where the spawning can happen instead of a biome, so we hope it works. In addition, the use of the over powered iron golem may be quite surprising for anyone stupid enough to attack it, and it may inspire pvpers to group up to kill it, assuming future rewards are good enough (like lore items for bastions or whatever).